removing duplicate values

+++ src/components/SearchedCity.js
@@ -25,9 +25,25 @@ const SearchedCity = (props) => {
     
 
     let storeId = () => {
-        ids.push(props.cityId);
-        localStorage.setItem('ids', JSON.stringify(ids));
-        props.close();
+        let exists = false;
+        if(ids.length > 0){
+            for(let i = 0; i < ids.length; i++){
+                if(ids[i] === props.cityId){
+                    exists = true;
+                    break;
+                }
+            }
+            if(!exists){
+                ids.push(props.cityId);
+                localStorage.setItem('ids', JSON.stringify(ids));
+            }
+            props.close();
+        }
+        else{
+            ids.push(props.cityId);
+            localStorage.setItem('ids', JSON.stringify(ids));
+            props.close();
+        }
     } 
 
     return (
